+++ integration-tests/udp/api_test.go
@@ -80,9 +80,9 @@ func TestSetTime(t *testing.T) {
 		Protocol: "udp",
 	}
 
-	dateTime := types.MustParseDateTime("2025-11-04 12:34:56")
+	datetime := types.MustParseDateTime("2025-11-04 12:34:56")
 
-	response, err := uhppoted.SetTime(u, controller, dateTime, timeout)
+	response, err := uhppoted.SetTime(u, controller, datetime, timeout)
 
 	if err != nil {
 		t.Fatalf("%v", err)