+++ src/uhppoted/responses/generated.go
@@ -363,6 +363,13 @@ type SetAntiPassback struct {
 	Ok         bool   `json:"ok"`
 }
 
+// Container struct for the response returned by a controller when setting
+// the first card configuration for a door.
+type SetFirstCard struct {
+	Controller uint32 `json:"controller"`
+	Ok         bool   `json:"ok"`
+}
+
 // Container struct for the response returned by a controller after restoring the manufacturer
 // default settings.
 type RestoreDefaultParameters struct {